:construction: Two new features approve Vacant and deny Vacant

+++ src/Routes/executive.routes.js
@@ -4,7 +4,7 @@ import executiveFunctions from "../Controllers/executive_mode";
 
 const router = Router();
 
-const { createMarket, getAllMarkets, getMarket, addVacant, deleteVacant, myExecutiveModes, registerExecutiveMode, addItem, editItem, deleteItem, addComment, addReply, editComment, editReply, deleteComment, deleteReply, addExecutiveAdmin } = executiveFunctions;
+const { createMarket, getAllMarkets, getMarket, addVacant, deleteVacant, myExecutiveModes, registerExecutiveMode, addItem, editItem, deleteItem, addComment, addReply, editComment, editReply, deleteComment, deleteReply, addExecutiveAdmin, deleteExecutiveAdmin } = executiveFunctions;
 
 //// MARKET ROUTES
 router.post("/createMarket", createMarket);
@@ -39,6 +39,8 @@ router.post("/deleteComment", deleteComment);
 router.post("/deleteReply", deleteReply);
 //addExecutiveAdmin
 router.post("/addExecutiveAdmin", addExecutiveAdmin);
+//deleteExecutiveAdmin
+router.post("/deleteExecutiveAdmin", deleteExecutiveAdmin);
